+++ spec/internal/app/models/event.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+class Event < ApplicationRecord
+  include Avocado::Event
+end